Fix misaligned raster extent after reprojection
