+++ raster/src/main/scala/geotrellis/raster/reproject/ReprojectRasterExtent.scala
@@ -31,60 +31,28 @@ import spire.syntax.cfor._
 object ReprojectRasterExtent {
   import Reproject.Options
 
-  /*
-   * This function is used to suggest the extent and resolution
-   * appropriate given the indicated reprojection CRSs.  It walks
-   * the edges of the input file (approximately 20 sample points along each
-   * edge) transforming into output coordinates in order to get an extents box.
-   */
-  def reprojectExtent(ge: GridExtent, transform: Transform): Extent = {
-    val extent = ge.extent
-    val (cols, rows) = (extent.width / ge.cellwidth, extent.height / ge.cellheight)
-    val PIXEL_STEP = math.min(50.0, math.min(cols, rows)).toInt
-
-    // Find the threshold to densify the extent at.
-    val xThreshold = (cols / PIXEL_STEP) * ge.cellwidth
-    val yThreshold = (rows / PIXEL_STEP) * ge.cellheight
-    val threshold = math.min(xThreshold, yThreshold)
-
-    // Densify the extent to get a more accurate reprojection
-    val denseGeom = Polygon(Densifier.densify(extent.toPolygon.jtsGeom, threshold).asInstanceOf[org.locationtech.jts.geom.Polygon])
-    denseGeom.reproject(transform).envelope
-  }
-
   def apply(ge: GridExtent, transform: Transform, options: Options): GridExtent = {
     val extent = ge.extent
-    val newExtent = reprojectExtent(ge, transform)
-
-    val (transformedXmin, transformedYmax) = {
-      transform(extent.xmin, extent.ymax)
-    }
-
-    val (transformedXmax, transformedYmin) = {
-      transform(extent.xmax, extent.ymin)
-    }
+    val newExtent = extent.reprojectAsPolygon(transform, 0.001).envelope
 
     options.parentGridExtent match {
       case Some(parentGridExtent) =>
-        parentGridExtent.createAlignedGridExtent(Extent(transformedXmin, transformedYmin, transformedXmax, transformedYmax))
+        parentGridExtent.createAlignedGridExtent(newExtent)
       case None =>
         val (pixelSizeX, pixelSizeY) =
           options.targetCellSize match {
             case Some(cellSize) =>
               (cellSize.width, cellSize.height)
             case None =>
-              val distance = (transformedXmin, transformedYmax).distance((transformedXmax, transformedYmin))
+              val distance = newExtent.northWest.distance(newExtent.southEast)
               val cols = ge.extent.width / ge.cellwidth
               val rows = ge.extent.height / ge.cellheight
               val pixelSize = distance / math.sqrt(cols * cols + rows * rows)
               (pixelSize, pixelSize)
           }
 
-        val newColsDouble = newExtent.width / pixelSizeX
-        val newRowsDouble = newExtent.height / pixelSizeY
-
-        val newCols = (newColsDouble + 0.5).toInt//.toLong
-        val newRows = (newRowsDouble + 0.5).toInt//.toLong
+        val newCols = (newExtent.width / pixelSizeX + 0.5).toInt//.toLong
+        val newRows = (newExtent.height / pixelSizeY + 0.5).toInt//.toLong
 
         //Adjust the extent to match the pixel size.
         val adjustedExtent = Extent(newExtent.xmin, newExtent.ymax - (pixelSizeY*newRows), newExtent.xmin + (pixelSizeX*newCols), newExtent.ymax)
